feat: setting up Spring Modulith and config the project´s folder structure
