+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/Barbershops/Domain/Model/BarberShops.java
@@ -1,0 +1,34 @@
+package multi_tenant_back.tenant_api.Barbershops.Domain.Model;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import multi_tenant_back.tenant_api.Barbershops.Infrastructure.BarberShopsDto;
+import multi_tenant_back.tenant_api.CitiesStates.Domain.Model.CitiesModel;
+
+import java.util.UUID;
+
+@Entity
+@Table(name = "barber_shops")
+@Getter
+@Setter
+@RequiredArgsConstructor
+public class BarberShops {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+    private String name;
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "city_id")
+    private CitiesModel city;
+    private String address;
+
+    public BarberShops(BarberShopsDto dto){
+        this.city = dto.getCity();
+        this.name = dto.getName();
+        this.address = dto.getAddress();
+    }
+}
