+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/Barbershops/package-info.java
@@ -1,0 +1,8 @@
+@ApplicationModule(
+        id = "barbershops",
+        displayName = "Barbershops",
+        type = ApplicationModule.Type.OPEN
+)
+package multi_tenant_back.tenant_api.Barbershops;
+
+import org.springframework.modulith.ApplicationModule;
