+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/CitiesStates/Domain/Model/StateCitiesDTO.java
@@ -1,0 +1,14 @@
+package multi_tenant_back.tenant_api.CitiesStates.Domain.Model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class StateCitiesDTO {
+    private int id;
+    @JsonProperty("departament")
+    private String state;
+    private List<String> cities;
+}
