--- tenant-api/src/main/java/multi_tenant_back/tenant_api/CitiesStates/model/CitiesModel.java
@@ -1,20 +1,0 @@
-package multi_tenant_back.tenant_api.CitiesStates.model;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "cities")
-@Getter
-@Setter
-public class CitiesModel{
-    @Id
-    @GeneratedValue(strategy =  GenerationType.AUTO)
-    private Long id;
-    private String name;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "state_id")
-    private StateModel state;
-}
