+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/CitiesStates/package-info.java
@@ -1,0 +1,8 @@
+@org.springframework.modulith.ApplicationModule(
+        id                  = "CitiesStates",
+        displayName         = "Cities & States",
+        type = ApplicationModule.Type.OPEN
+)
+package multi_tenant_back.tenant_api.CitiesStates;
+
+import org.springframework.modulith.ApplicationModule;
