--- tenant-api/src/main/java/multi_tenant_back/tenant_api/CitiesStates/repository/CitiesRepository.java
@@ -1,7 +1,0 @@
-package multi_tenant_back.tenant_api.CitiesStates.repository;
-
-import multi_tenant_back.tenant_api.CitiesStates.model.CitiesModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CitiesRepository extends JpaRepository<CitiesModel,Long> {
-}
