+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/People/package-info.java
@@ -1,0 +1,8 @@
+@org.springframework.modulith.ApplicationModule(
+        id = "people",
+        displayName = "People",
+        type = ApplicationModule.Type.OPEN
+)
+package multi_tenant_back.tenant_api.People;
+
+import org.springframework.modulith.ApplicationModule;
