--- tenant-api/src/main/java/multi_tenant_back/tenant_api/PhoneNumbersTenant/Configuration/PhoneNumberConfiguration.java
@@ -1,6 +1,0 @@
-package multi_tenant_back.tenant_api.PhoneNumbersTenant.Configuration;
-
-public final class PhoneNumberConfiguration {
-    public static final Integer LENGTH_DIGITS = 11;
-    public static final String PHONE_PREFIX = "+57";
-}
