+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/PhoneNumbersTenant/package-info.java
@@ -1,0 +1,8 @@
+@org.springframework.modulith.ApplicationModule(
+        id          = "phone_tenant",
+        displayName = "Shared Phone Tenants",
+        type = ApplicationModule.Type.OPEN
+
+)
+package multi_tenant_back.tenant_api.PhoneNumbersTenant;
+import org.springframework.modulith.ApplicationModule;
