+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/Roles/Domain/Model/RolesModel.java
@@ -1,0 +1,16 @@
+package multi_tenant_back.tenant_api.Roles.Domain.Model;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "roles")
+@Getter
+@Setter
+public class RolesModel {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Integer id;
+    private String name;
+}
