+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/Roles/Domain/Repository/RolesRepository.java
@@ -1,0 +1,7 @@
+package multi_tenant_back.tenant_api.Roles.Domain.Repository;
+
+import multi_tenant_back.tenant_api.Roles.Domain.Model.RolesModel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RolesRepository extends JpaRepository<RolesModel, Integer> {
+}
