--- tenant-api/src/main/java/multi_tenant_back/tenant_api/Roles/RolesModel.java
@@ -1,16 +1,0 @@
-package multi_tenant_back.tenant_api.Roles;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "roles")
-@Getter
-@Setter
-public class RolesModel {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    private String name;
-}
