--- tenant-api/src/main/java/multi_tenant_back/tenant_api/Roles/RolesRepository.java
@@ -1,6 +1,0 @@
-package multi_tenant_back.tenant_api.Roles;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RolesRepository extends JpaRepository<RolesModel, Integer> {
-}
