+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/Roles/package-info.java
@@ -1,0 +1,7 @@
+@org.springframework.modulith.ApplicationModule(
+        id          = "roles",
+        displayName = "Shared Roles",
+        type = ApplicationModule.Type.OPEN
+)
+package multi_tenant_back.tenant_api.Roles;
+import org.springframework.modulith.ApplicationModule;
