+++ tenant-api/src/main/java/multi_tenant_back/tenant_api/utils/package-info.java
@@ -1,0 +1,8 @@
+@org.springframework.modulith.ApplicationModule(
+        id          = "utils",
+        displayName = "Shared Utils",
+        type = ApplicationModule.Type.OPEN
+)
+
+package multi_tenant_back.tenant_api.utils;
+import org.springframework.modulith.ApplicationModule;
